+++ tests/testthat/test-target_sda.R
@@ -275,9 +275,9 @@ test_that("with no matching data warns", {
       fake_matched(),
       fake_ald(),
       bad_scenario,
-      region_isos = region_isos_demo
+      region_isos = region_isos_stable
       ),
-    "no scenario"
+    class = "no_match"
   )
 })
 
@@ -738,4 +738,29 @@ test_that("argument `weight_emission_factor` outputs correctly with known input 
     out$`boral cement`$emission_factor_value,
     ald$`boral cement`$emission_factor
   )
+})
+
+test_that("outputs empty tibble for sectors in `scenario` and `ald` but not
+          `data` (#390)", {
+
+  ald <- fake_ald(
+    sector = c("cement", "steel")
+  )
+
+  scenario <- fake_co2_scenario(
+    sector = rep(c("cement", "steel"), each = 2),
+    year = rep(c(2025, 2026), 2),
+    emission_factor = rep(c(1, 2), 2)
+  )
+
+  out <- target_sda(
+    fake_matched(sector_ald = "cement"),
+    ald,
+    scenario,
+    region_isos = region_isos_stable
+  )
+
+  out_steel <- filter(out, sector == "steel")
+
+  expect_equal(nrow(out_steel), 0L)
 })